The Files "DataTypeTesting.java" and "Test2.java" have been removed.
"BarSeries.java" and "DoingTheMath.java" have been worked on and are
close to being finished (both classes have problems with the DataType
"date", and seeing that SWTChart can only display doubles, I doubt that
the DataType "date" can be displayed at all with this library...). Next,
I will be working on a Box-plot method/class.